Code Climate cleanup   -add rubocop to rake task
