+++ Rakefile
@@ -1,8 +1,11 @@
-require "bundler/gem_tasks"
+require 'bundler/gem_tasks'
 require 'rspec/core/rake_task'
 
 # Setup new RSpec for rake task
 RSpec::Core::RakeTask.new(:spec)
 
-# make default task to run specs
-task :default => :spec
+require 'rubocop/rake_task'
+RuboCop::RakeTask.new
+
+task default: [:spec]
+task test: :spec
