+++ lib/seat_geek/taxonomies/concert.rb
@@ -1,6 +1,6 @@
 module SeatGeek
   module Taxonomies
-    class Concert
+    class Concert < ChildTree
       def initialize(child_taxonomies:)
         @taxonomies = child_taxonomies
       end
@@ -15,18 +15,6 @@ module SeatGeek
       end
 
       attr_accessor :taxonomies, :parent_id
-
-      private
-
-      def build_child_tree
-        parent_object = taxonomies.detect { |taxonomy| taxonomy['id'] == parent_id }
-
-        child_taxonomies = [parent_object]
-        taxonomies.each do |taxonomy|
-          child_taxonomies << taxonomy if taxonomy['parent_id'] == parent_id
-        end
-        child_taxonomies
-      end
     end
   end
 end
