+++ lib/seat_geek/taxonomies/sports.rb
@@ -1,6 +1,6 @@
 module SeatGeek
   module Taxonomies
-    class Sports
+    class Sports < ChildTree
       def initialize(child_taxonomies:)
         @taxonomies = child_taxonomies
       end
@@ -25,7 +25,7 @@ module SeatGeek
       end
 
       def football
-        @parent_id = 1020000
+       @parent_id = 1020000
         build_tree
       end
 
@@ -68,17 +68,6 @@ module SeatGeek
 
       attr_accessor :parent_id, :taxonomies
 
-      def build_child_tree
-        # parent_ids.each do |parent_id|
-        parent_object = taxonomies.detect { |taxonomy| taxonomy['id'] == parent_id }
-
-        child_taxonomies = [parent_object]
-        taxonomies.each do |taxonomy|
-          child_taxonomies << taxonomy if taxonomy['parent_id'] == parent_id
-        end
-        child_taxonomies
-        # end
-      end
     end
   end
 end
