+++ lib/seat_geek/version.rb
@@ -1,3 +1,3 @@
 module SeatGeek
-  VERSION = "0.14.0"
+  VERSION = "0.14.1"
 end
